requirements.txt
Django>=4.2,<5
python-dotenv
requests
PyPDF2
python-docx
pymupdf
sentence-transformers
numpy
chromadb
langchain
langchain-community
langchain-ollama

